Caseworker: Remove Authentication middleware
This removes auth middleware and auth backends because they were the only thing
that relied on a database.

I have moved all of the "user" state on to request.session but will continue to
investigate a good way of abstracting this. Possibly by adding a new middleware
that puts something on `request.user` but constructs that object from the
session rather than the database.