LTD-507 Control List Entries cleared when adding new product to app
This fixes a bug with the serialisation of a new good when being added to an
application, there was an issue that if the user added a firearm certificate
then the control list entries would be lost. Fixing that bug in `lite-api`
also uncovered another bug in that only the first control list entry was
being sent to lite-api instead of the full list.

This commit makes sure the full list of `control_list_entries` is saved to the
session so it can be played back later.