Design changes for the way flags are added and edited in routing rules
A routing rule can be applied if the case contain certain flags and
does not contain some other flags. The UI allows setting up these two
conditions and had to make sure a flag is included in either of the
condition but not both. This change updates the design and fixes some
of the inconsistencies in previous implementation.

All of the changes are in javascript mostly to show/hide controls,
enable/disable and check/uncheck flags.