LTD-795 Update Tokenfield monkeypatch
LITE uses a library called tokenfield, this was previously vendored; recently it
was moved to a npm managed library. The version from npm was newer than our
vendored one and the 'monkeypatch' introduced to make it remove `[]` from the end of
input fields stopped working.

This commit fixes that 'monkeypatch' in a slightly more clean way by overriding
the TokenField class.

It also moves the TokenField into the js bundle instead of only loading it on
review-good-base template.