Fix denials upload failure
Upload currently failing with the error that "File was not opened in read mode".
We are using FILE_UPLOAD_HANDLERS which defaults to S3FileUploadHandler.
This is creating a S3Boto3StorageFile instance in write mode so when
we try to read it from this, it is complaining that it is not opened in read mode.

Fix this by reading it from the s3 object that is uploaded.