Fix display of Control list entries on Product detail page
Control list entries are in two places in GoodOnApplication instance, before
a good is reviewed they could be different.
if is_good_controlled is None, then it is not yet reviewed by the Case officer so we
display the entries from the good.
if is_good_controlled is not None (either Yes or No) then it means the good is reviewed
and the Case officer applied the correct CLC entry so display from good_on_application.
Add unit test to verify the values displayed.