Display Section5 details on product detail page in all cases
When user adds a product that is covered under section5 for the first time then
we provide a form to upload the certificate, thereafter if another product is
added under section5 the system identifies that there is a valid section5
certificate and skips the upload form. Because of this the product detail page
shows the certificate details as empty/None eventhough there is a valid certificate.
It cannot show this information because it is not saved in the firearm details and
have no access to the organisation documents (RFD, section1/2/5 certificates).

This is fixed by exposing these documents and displaying the details correctly if
there is a valid section5 certificate.