Enable UI tests on CI and skip all of them for now
Browser tests are now enabled on CI but it will only run those tests that are
not marked as "skip".

This commit also marks all the tests to be skipped as well for now except login and logout
test. This is because the tests are currently broken but we want to make the setup ready
hence everything is skipped. They will be unskipped one by one as we fix/rewrite the
required browser tests.