Provide GoodOnApplication id as key when reviewing the product
When there are multiple products on an application we select the ones to be
reviewed and click "Review" button. The review page is provided the list
of ids of the underlying "Good" instance in "GoodOnApplication".

This works ok if the same Good is added once but if the user used the same
"Good" to add multiple products to the application then we are always
selecting the first "GoodOnApplication" instance that matches with the given
Good id hence displaying incorrect quantity value in the review page.

Fix this by providing the GoodOnApplication id instead, that way we always
select the right instance and display correct values.

Update unit tests accordingly.