Fix product flags form to provide good id to assign flags correctly
This form allows user to assign specific flags to a product on application and it
receives a list of ids. These were product ids (Good model) but when we fixed an
issue with review goods form this has changed to ids of GoodOnApplication.

Because of this the backend is not able to find relevant object and throwing
'not found' error as response. Fix this by extracting the relevant Good ids
in the form.