Fix updating of ARS for a specific product on application
Annual report summary (ARS) is specific to a product on application. If the same product
is used on other application then updating ARS in one application was updating the
value in the other application also. This has been fixed but the fix was failing
if the same product is added twice in the same application.

This is because we provide the list of underlying product id (Good) to the api and
it retrieves all matching products on application (GoodOnApplication). In this case it
will get all products in that application (as same product is added twice) and updates
all entries. This PR fixes this by also providing the current product on app id to the
api which is used to limit the updates to a particular entry within the application.

Also fixes loading the initial data when multiple products are reviewed using wizard.