Fixup: Create custom field for picklist and avoid repeating js
Since the same pattern is being repeated multiple times a custom field
is created for fields that can be selected from picklists. It also
accepts certain attributes which are used in JS to retrieve
relevant picklists from the API.

The javascript also moved to a template to avoid repeating it for
each field, functions are parameterized so they work for all
picklist target and types.

The picklist picker dialog html also is made generic and is not
repeated for each picklist field.