Fix Edit advice to save the updated data correctly
When editing advice we currently use the previously retrieved advice
and use it as initial data to retrieve the form. This is fine but when
the user made changes and submitted, we are still using the same data
instead of using the POST data.
Because of this form is getting submitted even if the mandatory fields
are omitted and changes are not getting saved. Fix this by using the
POST data when the form is submitted.