Retain current case queues when sending it back after countersign refuse
When the countersigner doesn't agree with the recommendation the case is
sent back to advisor along with the queue it needs to go back but this
was unassigning the case from it's current set of queues. We only want
it to be unassigned from the countersigning queue and it fixes this.