Fix display of security grading value in case details summary
Currently when displaying security grading we are looking for pv_grading_details
which is actually a dict, we should be using is_pv_graded which is a string and
the values of all products are joined and displayed as a comma separated value.

If the product is security graded we should also display the actual grading
instead of yes/no so a follow up is required to display more details.