Update display of PV security grading details in case details section
In Case details side panel currently we only display yes/no to indicate
whether a product is pv graded or not in both individual product section
and in summary section below which is not very useful. This is updated to
show the classification level as well if a product is pv graded.

If none of the products are pv graded then that field is completely removed
from individual section and it is mentioned as 'No' in the summary.

If any of the product is pv graded then we should the classification and
mention it as 'Yes' in the summary section.

v2: Address review comments
 - Use Django yesno filter