Post advice for NLR items also
If the licence is being issued for an application and it contains
NLR items then we have to generate NLR letter besides the main licence
document. The NLR letter is only generated if there is a final advice
for NLR items also.

To keep it consistent with Advice1.0, post NLR advice for other
advice levels as well.