Check flags of all levels to determine if LU countersign required or not
Some of the LU countersign required are of flag level "destination" so while
determining whether LU countersign required or not check flags of all levels.
Currently we are only checking flags of "Case" level and thus determining
that countersigning is not required even though the flag is applied to the case.