Change submit button label to underline two step process
Users after submitting recommendation they have move the case forward so
that it is routed to correct team. If the case is not moved forward then it
will still remain in their queue. This is a two step process. Having the
label as just submit may cause confusion as it has already been submitted
to other teams so change the label to make it more clear.