Add view to edit the End-user undertaking document
This form has additional options such as whether the document is in english
and the document is on letterhead etc. User might have uploaded these documents
earlier but if he chooses to change those options then we need to ask those
questions accordingly hence this view is created as a wizard and depending on
the updated options we update the journey in the wizard.

If user uploads another document then we delete the existing document.
If user chooses not to upload then we allow them to continue without raising
any validation error.