Capture and display product differences detail from End-user
When uploading the end-user undertaking document we ask the user if there
are any differences in the products listed in the document and the products
in the application. Previously this was incorrectly associated as document
description but infact it is a separate piece of detail we are capturing
from the user.

Views are updated to capture and present this information and templates
for end-user summary and in other places are updated