Show EC3 banner if user hasn't provided these details when required
We ask End-user to provide EC3 document or the reason for not having it under
certain conditions. This is only asked if the products are originating from NI,
going to EU and the product type is firearms but not related to software or
technology.

Since user can complete different sections in different order, it is often not
possible to determine if EC3 is required or not. Eg user can fill end-user
details first following by adding products and specifying goods starting point etc.

For this reason once we determine user needs to provide EC3 details we display
a banner on the task list page asking the user to provide these details. Once
these details are entered the banner is not shown. Since user can always edit
these details the banner comes up and goes as and when it is applicable.

Both EC3 document and the missing reason are optional so when the banner is shown,
if user provides any of these details then it is removed.