Show all parties documents in one place
Case summary page has a new section to show End-user documents. This is
added as part of English translation document changes but currently we also
show these details in destinations table on the same page so the designs
are updated to show all parties documents in this new section and remove
it from the destinations table.