Add ui test that verifies enforcement check by importing the file
Enforcement checks are done on parties of an application. We first export
the data in an xml file and after the checks are performed another file
is imported which contains data to update the application based on the
checks.

In this test we generate the file to be imported based on the parties
data and once imported we assert that the application is updated or not.