Fix the PII checker and Black failing pre-commits
Click version 8.1.0 was released recently and it changed some of the
internals. This caused a new clean install of the PII checker tool
to fail with an ImportError exception in a transitive dependency.

For now we can pin the dependency to an older version.

Black released a new version that fixes the use of internals.