Add view to edit Firearm product type
Firearm product type is a basic characterstic and changing this means some of
the question already answered may become invalid and new question may be
applicable hence take the user again through the questions specific to the
updated type.

The questions that are common across all types are not asked again as they
can be edited from individual edit links.