Use the correct template for the wizard
One of the entry point for this wizard is when user edits the End-user
document availability option. That form is a single template view so it
uses a different template_name and the same is being used in the wizard
also which causes management form errors so explicity override again
to use the correct name in this wizard.