Remove I'm done button for LU users when case is on post circ queue
When adding post circulation notes LU users are clicking I'm done
inadvertently as it is available on that page. The problem with
doing this is that they are not finalising the case and also because
this is the last step and the case becomes queueless.

A long term solution is in the works but till then we are removing
this button for LU  when the case is on post-circulation queue.

If the case is on any other queue besides this one then there is no
issue in showing this button because it won't become queueless.