Fix I'm done button for LU users
We want to remove I'm done button for LU users if the case is only
on Licensing unit post circulation queue. The current code uses
queue alias to determine this but it is checking whether alias is
valid before taking this value. This condition is actually not
required because the problem with this is that alias is not defined
for all queues, eg if it is on team leader to countersign queue for
which the alias is not defined then the current code filters this
out and removes the button in this case also which is not correct.