Enable MOD-DI Direct and FCDO CPACC users to give their recommendation
The option to give recommendation is only available if the user's current
queue matches with the queue the case is currently on. We match this
based on queue alias.

As part of routing changes we added new queues so include their aliases
in this check so that users processing a case on these queues can give
their recommendation.

Also removed MOD-WECA aliases which we no longer need.