Add option to review potential trigger list products
In case of BEIS Nuclear before giving recommendation they need to assess products
and confirm whether they are on trigger list or not. So add an option to assess
products. Once they complete their assessment then option to give advice will
be available.

Currently assess products option does nothing which will be updated once we
add that form. Also this is shown even if there are no products to assess.
We will conditional enable this later.

A helper function is added to filter potential trigger list products from
all the products on the application.