Filter NLR products when re-issuing a licence
In case of re-issuing a licence, LU needs to finalise the application again.
We use a different form when re-issuing as opposed to initial licence issue
and this form currently is not filtering NLR products. It is not aware of
NLR products and showing them in this case.

When the form is submitted we attach those NLR products also to the licence.
They in turn get included in the licence pdf and also in the details sent to
HMRC which is not correct.

Update the form to account for NLR products and exclude them from including
on the licence.