Switch LU to use put method to update consolidated advice
When LU edits consolidated advice we don't want to delete advice objects
as it can result in deletion of Countersign objects (if exists) because
of cascade deletes hence we need to actually update them.

A new put method is added in API so switch LU to use this method when
editing consolidated advice.

Because it is required for LU countersigning work these changes are
restricted to LU only but can easily be expanded to other teams.