LTD-3407: pass `return_to` url through case assignment flow
When a user clicks the 'Allocate case' button from a case detail page,
we would like them to be returned to that case after they have completed
the journey to assign either a case officer or a case adviser.

This commit passes the `return_to` param from the button through the
different forms used.