Update test to check for countersign warning when LU caseworker edits recommendation
When a Case is sent back we show a warning that it has been sent back but
when caseworker edits recommendation, all previous countersignatures are
marked as invalid and previous warning is no longer relevant because it
will start countersign journey again.

In this case we display the usual warning that this case needs countersigning.

Update test to check for this warning in case of invalid countersignatures.