Use patch Case endpoint to update case status
We currently use an endpoint in applications app to update case status.
This works but it doesn't result in case.save() and as a result
doesn't trigger the required signals which is essential to running
flagging rules when Case status change.

Use the patch endpoint available in cases app instead which triggers
the signals correctly