Update cases url to include default queue
Currently when user selects 'Cases' menu option after switching Team and
default queue, it points to root url. It doesn't include the queue
even though we show cases from that particular queue.

This seems to be the cause of the flakiness in UI tests when the
Case filters are converted to use django-forms.

UI tests particularly the ones in give_advice.feature switches teams,
filters cases using reference code multiple times during the journey.

During this menu option 'Cases' is selected and 'Clear filters' is
clicked. In both cases user is taken to the current queue except that
in the case of clearing filters current queue_pk is included in the url.

In the filters form we have a property for queue_pk which is either
queue from kwargs or from the one saved in the session. Because of this
current queue that we get was different sometimes and the test was failing
as it couldn't find the Case in that queue. It was failing in the CI so
a bit difficult to trace the exact issue.

This change makes it consistent so that we get the current queue the
same in all cases. Tests are passing consistently with this change.