Consider NLR products when LU edits their recommendation
Currently when LU edits their recommendation (final-level) then we are
updating all products as either approve/refuse depending on their decision.
This includes NLR products also which means if application has any NLR products
they will also be recorded as approve/refuse after they edit recommendation.

This is a problem as it is final level advice because they are no longer
treated as NLR and they will be added to the licence document also.

Fix this by filtering NLR products and recording the type as NLR correctly
for all NLR products. We have existing tests so update them to include
NLR products.