Update help text to clarify the functionality to screen reader users
When assigning LU case officer typing into the input field filters the
options available for the users to select. This is not very clear for
screen reader users so as per suggested solution update help text so
that it clear for them before they start interacting with the input field.