Correct typo for data-module field to focus on error summary
In case of form errors keyboard focus should be taken to the error summary
but because of a typo to the value of data-module field it is incorrectly
taken to other elements on the page. This causes issues for screen reader
users as they are not made aware of the error(s).