Fix contrast of back links
Back links are currently displayed in white color against a dark background.
When they are cllicked the focus color is yellow and it makes it difficult
to read as the text is white. To improve the contrast the text color is
changed to black when the link is in focus.