Group search results based on Product CLE
When product search results are displayed we show all the cases
where this particular product is used.

Most of the time a particular product is rated with a particular CLE
so all of the entries for remaining cases will have the same rating.
Case advisers are only interested in distinct rating so in order to
avoid duplicating the same for each application they are grouped by
rating and one entry is show in the main section. Remaining hits are
hidden by default and they can examine if required.

It is possible that the same product may be rated different depending
on other attributes on that application so in this case we show distinct
possible ratings first and the rest of them in a separate group.