Update the way of checking for expected fields in test
Previously we were checking for exact matching of list of fields.
This works but breaks if we add new fields in the API. We are going
to add more fields so update it to check it is a subset.

Also parse html and compare the contents of search results table with
the data from fixtures. This will catch issues if we are using
incorrect fields as that value will be empty in the template.