Remove option to search for products in Product details page
When you view full product details we also show where that paricular
product is used in other applications in "Cases" tab.

This tab also provides an option to search for products but it points
to initial proof of concept product search. We now have a dedicated
page for product search so this is no longer required.