Update product search test to check for presence of expected hits
As this is an E2E test, things are driven from the UI. We also have
many other tests running which would've created more products and
their names, ratings may match with some of the values used in
this test so when running search scenarios that will affect the
results if we are looking for exact counts.

We cannot clear other products before this test is run, hence
update the test assertions to look for the presence of expected
suggestions, hits rather than exact counts.