Remove search proof of concept, spire search and dependent packages
Some of the packages (directory-api-client) require Django version
slightly lower than the LTS version (4.2.8).

These are mainly required for search proof-of-concept app and also
an app that searches SPIRE data from a snapshot. We now have a
Product search dedicated for TAU so we no longer need this poc
hence remove these apps and dependent packages. There is a plan
to clean this up anyway so remove during this upgrade.