Update attribute used to get class name
In the upgraded URLResolver 'func.name' is not giving the class name
which is used to set up current_step name in the wizard.

Because of this we are getting management form errors when this
step is submitted, update the attribute used to get the class name
to fix this.