Remove manage_cases.feature tests
This test manually updates Case status and ensure that this has been
updated. This test is not relevant because we don't change status
manually. Case status changes when the routing rules are applied
as it progress through various teams.

Also we need to assign a case advisor to change the status first
and those steps are missing in this test.