Remove tests in assign_users_to_queue feature file
These tests are not relevant as creating new queues is managed using
migrations. These tests create a queue and try to assign users.

Before any action is performed user needs to assign the case to himself
and it is unassigned when the case is progressed by routing rules.
So both of these actions are covered multiple times as part of other tests.