Remove give_advice_form lite-form and associated view
This form is part of Advice1.0 which is used for giving user recommendation
on a Case however it is no longer used as it is replaced by Advice2.0 work.

All of the advice related forms are under a new app called 'advice'.

Removing the form, associated view and unit tests.