Add UI test to check for details in the SIEL licence document
This test creates a SIEL application, progresses it through all the
teams as per the process and issues a licence at the end.

In the generated licence document it verifies the details match with
the expected values.

We don't have any test that currently checks the document and this test
fills this gap. It checks the following,
 - licence contains products that needed licence and in correct order
 - licence doesn't contain NLR products
 - licence contains proviso (if any)
 - product details (Name, part number, serial numbers (if applicable), CLEs etc)
 - Party details
 - Export type

Some of the products are NLR so it generates NLR letter and checks that
the letter contains details of correct NLR products.