Do not use licence to finalise goods on the application
This can cause an issue if we are re-issuing a licence because one
case where a licence is re-issued is if application is reopened
and edited to remove some of the existing goods. If licence is
used in this case then we won't consider any of the new goods
that are added to application. To fix this retrieve current
goods on application when finalising them.