Add additional info questions for F680 application
This section uses lite-forms. It is currently using a SummaryListView which
is not working so changed it to use MultiFormView.

The issue with SummaryListView is that after submitting to the last step
we are going back to first step again instead of redirecting.

Field names updated to match the model in API.