Provide option to create a copy of application when performing major edits
With major edits application need to be resubmitted. This can interfere with
the existing state of the application so it is better to create a copy of
existing application and then make required amendments. Also resubmitting
starts the processing from the beginning and can take more time.

If user decided to make major changes then provide an option to confirm creating
a copy of application.