Skip the E2E test for amending an application by copy
Opening an application for editing by copying an existing application
is allowed for configurable list of Exporters. This behaviour is
not possible to imitate in E2E test as it is a runtime check and
the Exporter Id changes for each iteration.

As an alternative we added this test assuming it is available for all
exporters and skipping the test for now.

Once this feature is opened for all then we will enable this test.