Fix licence manager issue where he can edit countersignature from another queue
When presenting option to countersign or edit countersignature we check if
the current user has countersigned on this case. Here we check the
countersign advice on both countersigning queues so if this user already
countersigned on licencing manager queue and try to do the same from
senior licencing manager queue then we see that he has countersigned and
present with edit option.

Fix this by checking the advice for the specific queue.