Update and add new tests for cases in queue view
Existing tests are for a specific column in the queue view. The issue
with this is that we have different case types so each test checking
the column for all cases types. It is also not clear which case type
they are referencing in these tests as we currently refer as
first case, second case etc.

To make it better new tests are added for a specific case that checks
all columns applicable for that case type.