Update F680 forms to support selecting entities and recommendation type
When selecting the entities we can also choose the recommendation type.
This will helps us pull either provisos or refusal reasons in the next
form as we know the recommendation type.

If the team has not defined any provisos/refusal reasons then we
display a basic form with text field where they can enter details.